+++ src/main/java/de/samply/app/ProjectManagerApplication.java
@@ -1,5 +1,6 @@
 package de.samply.app;
 
+import net.javacrumbs.shedlock.spring.annotation.EnableSchedulerLock;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.persistence.autoconfigure.EntityScan;
@@ -12,6 +13,7 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EntityScan(basePackages = {"de.samply.db.model"})
 @EnableJpaRepositories("de.samply.db.repository")
 @EnableScheduling
+@EnableSchedulerLock(defaultLockAtMostFor = ProjectManagerConst.SHED_LOCK_MAX_TIME_SV)
 public class ProjectManagerApplication {
 
     static void main(String[] args) {