+++ src/main/java/de/samply/query/QueryEventService.java
@@ -1,6 +1,8 @@
 package de.samply.query;
 
+import de.samply.exporter.ExporterJob;
 import de.samply.project.ProjectBridgeheadService;
+import de.samply.project.SendQueryToBridgeheadEvent;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
@@ -9,9 +11,14 @@ import org.springframework.transaction.event.TransactionalEventListener;
 public class QueryEventService {
 
     private final ProjectBridgeheadService projectBridgeheadService;
+    private final ExporterJob exporterJob;
 
-    public QueryEventService(ProjectBridgeheadService projectBridgeheadService) {
+
+    public QueryEventService(
+            ProjectBridgeheadService projectBridgeheadService,
+            ExporterJob exporterJob) {
         this.projectBridgeheadService = projectBridgeheadService;
+        this.exporterJob = exporterJob;
     }
 
     /**
@@ -23,5 +30,15 @@ public class QueryEventService {
     public void onQueryChanged(QueryChangedEvent event) {
         projectBridgeheadService.handleChange(event.queryId());
     }
+
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    public void onQuerySentToBridgehead(@SuppressWarnings("unused") SendQueryToBridgeheadEvent event) throws InterruptedException {
+        // Execute twice to have enough time to check if the query has arrived to the bridgehead
+        for (int i = 0; i < 2; i++) {
+            exporterJob.checkExports();
+            Thread.sleep(5000);
+        }
+    }
+
 
 }